@@ -84,7 +84,7 @@
   },
   "devDependencies": {
     "@jest/test-sequencer": "^24.8.0",
-    "electron": "8.2.5",
+    "electron": "10.1.3",
     "electron-builder": "^22.7.0",
     "electron-builder-notarize": "^1.1.2",
     "eslint": "^6.8.0",
