@@ -16,17 +16,17 @@
     "@budibase/client": "^0.1.25",
     "@rollup/plugin-commonjs": "^11.1.0",
     "lodash": "^4.17.15",
-    "rollup": "^1.11.0",
+    "rollup": "^2.11.2",
     "rollup-plugin-commonjs": "^10.0.2",
     "rollup-plugin-json": "^4.0.0",
     "rollup-plugin-livereload": "^1.0.1",
     "rollup-plugin-node-resolve": "^5.0.0",
     "rollup-plugin-postcss": "^3.1.5",
-    "rollup-plugin-svelte": "^5.0.0",
-    "rollup-plugin-terser": "^5.1.1",
+    "rollup-plugin-svelte": "^5.0.3",
+    "rollup-plugin-terser": "^7.0.2",
     "shortid": "^2.2.15",
     "sirv-cli": "^0.4.4",
-    "svelte": "^3.12.1"
+    "svelte": "^3.29.0"
   },
   "keywords": [
     "svelte"
@@ -36,7 +36,7 @@
   "gitHead": "284cceb9b703c38566c6e6363c022f79a08d5691",
   "dependencies": {
     "@beyonk/svelte-googlemaps": "^2.2.0",
-    "@budibase/bbui": "^1.34.2",
+    "@budibase/bbui": "^1.41.0",
     "@fortawesome/fontawesome-free": "^5.14.0",
     "britecharts": "^2.16.1",
     "d3-selection": "^1.4.2",
