@@ -15,8 +15,12 @@ context("Create a automation", () => {
 
     cy.contains("automate").click()
     cy.contains("Create New Automation").click()
-    cy.get("input").type("Add Record")
-    cy.contains("Save").click()
+    cy.get(".modal").within(() => {
+      cy.get("input").type("Add Record")
+      cy.get(".buttons")
+        .contains("Create")
+        .click()
+    })
 
     // Add trigger
     cy.get("[data-cy=add-automation-component]").click()
@@ -46,7 +50,6 @@ context("Create a automation", () => {
 
     // Activate Automation
     cy.get("[data-cy=activate-automation]").click()
-    cy.contains("Add Record").should("be.visible")
     cy.get(".stop-button.highlighted").should("be.visible")
   })
 
