@@ -1,24 +1,12 @@
 <script>
-  import Modal from "svelte-simple-modal"
-  import { notifier } from "builderStore/store/notifications"
-  import { onMount, getContext } from "svelte"
-  import { backendUiStore, automationStore } from "builderStore"
+  import { onMount } from "svelte"
+  import { automationStore } from "builderStore"
   import CreateAutomationModal from "./CreateAutomationModal.svelte"
-  import { Button } from "@budibase/bbui"
+  import { Button, Modal } from "@budibase/bbui"
 
-  const { open, close } = getContext("simple-modal")
+  let modal
 
   $: selectedAutomationId = $automationStore.selectedAutomation?.automation?._id
-
-  function newAutomation() {
-    open(
-      CreateAutomationModal,
-      {
-        onClosed: close,
-      },
-      { styleContent: { padding: "0" } }
-    )
-  }
 
   onMount(() => {
     automationStore.actions.fetch()
@@ -26,7 +14,7 @@
 </script>
 
 <section>
-  <Button purple wide on:click={newAutomation}>Create New Automation</Button>
+  <Button primary wide on:click={modal.show}>Create New Automation</Button>
   <ul>
     {#each $automationStore.automations as automation}
       <li
@@ -39,6 +27,9 @@
     {/each}
   </ul>
 </section>
+<Modal bind:this={modal}>
+  <CreateAutomationModal />
+</Modal>
 
 <style>
   section {
@@ -57,7 +48,7 @@
     color: var(--grey-6);
   }
   i.live {
-    color: var(--purple);
+    color: var(--ink);
   }
 
   li {
