@@ -1,11 +1,8 @@
 <script>
   import { store, backendUiStore, automationStore } from "builderStore"
   import { notifier } from "builderStore/store/notifications"
-  import ActionButton from "components/common/ActionButton.svelte"
-  import { Input } from "@budibase/bbui"
+  import { Input, ModalContent } from "@budibase/bbui"
   import analytics from "analytics"
-
-  export let onClosed
 
   let name
 
@@ -13,72 +10,40 @@
   $: instanceId = $backendUiStore.selectedDatabase._id
   $: appId = $store.appId
 
+  function sleep(ms) {
+    return new Promise(resolve => {
+      setTimeout(resolve, ms)
+    })
+  }
+
   async function createAutomation() {
     await automationStore.actions.create({
       name,
       instanceId,
     })
-    onClosed()
     notifier.success(`Automation ${name} created.`)
     analytics.captureEvent("Automation Created", { name })
   }
 </script>
 
-<div class="container">
-  <header>
-    <i class="ri-stackshare-line" />
-    Create Automation
-  </header>
-  <div class="content">
-    <Input bind:value={name} label="Name" />
-  </div>
-  <footer>
-    <a href="https://docs.budibase.com">
+<ModalContent
+  title="Create Automation"
+  confirmText="Create"
+  onConfirm={createAutomation}
+  disabled={!valid}>
+  <Input bind:value={name} label="Name" />
+  <div slot="footer">
+    <a
+      target="_blank"
+      href="https://docs.budibase.com/automate/introduction-to-automate">
       <i class="ri-information-line" />
       <span>Learn about automations</span>
     </a>
-    <ActionButton secondary on:click={onClosed}>Cancel</ActionButton>
-    <ActionButton disabled={!valid} on:click={createAutomation}>
-      Save
-    </ActionButton>
-  </footer>
-</div>
+  </div>
+</ModalContent>
 
 <style>
-  .container {
-    padding: var(--spacing-xl);
-  }
-
-  header {
-    font-size: var(--font-size-xl);
-    color: var(--ink);
-    font-weight: bold;
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    align-items: center;
-  }
-  header i {
-    margin-right: var(--spacing-m);
-    font-size: 20px;
-    background: var(--purple);
-    color: var(--white);
-    padding: var(--spacing-s);
-    border-radius: var(--border-radius-m);
-    display: inline-block;
-  }
-
-  .content {
-    padding: var(--spacing-xl) 0;
-  }
-
-  footer {
-    display: grid;
-    grid-auto-flow: column;
-    grid-gap: var(--spacing-m);
-    grid-auto-columns: 3fr 1fr 1fr;
-  }
-  footer a {
+  a {
     color: var(--ink);
     font-size: 14px;
     vertical-align: middle;
@@ -86,10 +51,10 @@
     align-items: center;
     text-decoration: none;
   }
-  footer a span {
+  a span {
     text-decoration: underline;
   }
-  footer i {
+  i {
     font-size: 20px;
     margin-right: var(--spacing-m);
     text-decoration: none;
