@@ -1,5 +1,6 @@
 <script>
   import { join } from "lodash/fp"
+  import { TextArea, Label } from "@budibase/bbui"
 
   export let values
   export let label
@@ -15,35 +16,12 @@
   $: valuesText = join("\n")(values)
 </script>
 
-<div class="margin">
-  <label class="label">{label}</label>
-
-  <textarea value={valuesText} on:change={inputChanged} />
-
+<div class="container">
+  <TextArea {label} value={valuesText} thin on:change={inputChanged} />
 </div>
 
 <style>
-  .margin {
-    margin-bottom: 16px;
-    display: grid;
-  }
-  .label {
-    font-size: 14px;
-    font-weight: 500;
-    margin-bottom: 8px;
-  }
-  textarea {
-    font-size: 14px;
-    height: 200px;
-    width: 100%;
-    border-radius: 5px;
-    border: none;
-    cursor: text;
-    background: var(--grey-2);
-    padding: 12px;
-    -moz-box-sizing: border-box;
-    -webkit-box-sizing: border-box;
-    box-sizing: border-box;
-    font-family: Inter;
+  .container :global(textarea) {
+    min-height: 100px;
   }
 </style>
