@@ -1,32 +1,17 @@
 <script>
-  import Modal from "./Modal.svelte"
+  import SettingsModal from "./SettingsModal.svelte"
   import { SettingsIcon } from "components/common/Icons/"
-  import { getContext } from "svelte"
-  import { isActive, goto, layout } from "@sveltech/routify"
+  import { Modal } from "@budibase/bbui"
 
-  // Handle create app modal
-  const { open } = getContext("simple-modal")
-
-  const showSettingsModal = () => {
-    open(
-      Modal,
-      {
-        name: "Placeholder App Name",
-        description: "This is a hardcoded description that needs to change",
-      },
-      {
-        closeButton: false,
-        closeOnEsc: true,
-        styleContent: { padding: 0 },
-        closeOnOuterClick: true,
-      }
-    )
-  }
+  let modal
 </script>
 
-<span class="topnavitemright settings" on:click={showSettingsModal}>
+<span class="topnavitemright settings" on:click={modal.show}>
   <SettingsIcon />
 </span>
+<Modal bind:this={modal} width="600px">
+  <SettingsModal />
+</Modal>
 
 <style>
   span:first-letter {
@@ -35,8 +20,7 @@
   .topnavitemright {
     cursor: pointer;
     color: var(--grey-7);
-    margin: 0px 20px 0px 0px;
-    padding-top: 4px;
+    margin: 0 20px 0 0;
     font-weight: 500;
     font-size: 1rem;
     height: 100%;
