@@ -1,36 +1,29 @@
 <script>
-  import Button from "components/common/Button.svelte"
+  import { TextButton } from "@budibase/bbui"
+  import { Heading } from "@budibase/bbui"
+  import { Spacer } from "@budibase/bbui"
   export let name, _id
 </script>
 
 <div class="apps-card">
-  <h3 class="app-title">{name}</h3>
+  <Heading small black>{name}</Heading>
+  <Spacer medium />
   <div class="card-footer">
-    <a href={`/_builder/${_id}`} class="app-button">Open {name}</a>
+    <TextButton text medium blue href="/_builder/{_id}">
+      Open {name} →
+    </TextButton>
   </div>
 </div>
 
 <style>
   .apps-card {
     background-color: var(--white);
-    padding: var(--spacing-xl);
+    padding: var(--spacing-xl) var(--spacing-xl) var(--spacing-xl)
+      var(--spacing-xl);
     max-width: 300px;
     max-height: 150px;
     border-radius: var(--border-radius-m);
     border: var(--border-dark);
-  }
-
-  .app-button:hover {
-    background-color: var(--white);
-    color: var(--black);
-    text-decoration: none;
-  }
-
-  .app-title {
-    font-size: var(--font-size-l);
-    font-weight: 600;
-    color: var(--ink);
-    text-transform: capitalize;
   }
 
   .card-footer {
@@ -38,23 +31,5 @@
     flex-direction: row;
     align-items: baseline;
     justify-content: space-between;
-  }
-
-  .app-button {
-    align-items: center;
-    display: flex;
-    background-color: var(--ink);
-    color: var(--white);
-    border: 1.5px var(--ink) solid;
-    width: 100%;
-    justify-content: center;
-    padding: 8px 16px;
-    border-radius: var(--border-radius-s);
-    font-size: var(--font-size-xs);
-    font-weight: 500;
-    cursor: pointer;
-    transition: all 0.2s;
-    box-sizing: border-box;
-    font-family: var(--font-sans);
   }
 </style>
