@@ -1,20 +1,22 @@
 <script>
-  import { Input, Heading, Body } from "@budibase/bbui"
+  import { Label, Heading, Input } from "@budibase/bbui"
   export let validationErrors
   export let template
 
   let blurred = { appName: false }
 </script>
 
-{#if template}
-  <Heading small black>Selected Template</Heading>
-  <Body>{template.name}</Body>
-{/if}
 <h2>Create your web app</h2>
 <div class="container">
+  {#if template}
+    <div class="template">
+      <Label extraSmall grey>Selected Template</Label>
+      <Heading small>{template.name}</Heading>
+    </div>
+  {/if}
   <Input
     on:input={() => (blurred.appName = true)}
-    label="Web app name"
+    label="Web App Name"
     name="applicationName"
     placeholder="Enter name of your web application"
     type="name"
@@ -24,6 +26,11 @@
 <style>
   .container {
     display: grid;
-    grid-gap: 40px;
+    grid-gap: var(--spacing-xl);
+  }
+
+  .template :global(label) {
+    /*  Fix layout due to LH 0 on heading */
+    margin-bottom: 16px;
   }
 </style>
