@@ -1,5 +1,5 @@
 <script>
-  import { Button, Heading, Body } from "@budibase/bbui"
+  import { Button, Heading, Body, Spacer } from "@budibase/bbui"
   import AppCard from "./AppCard.svelte"
   import Spinner from "components/common/Spinner.svelte"
   import api from "builderStore/api"
@@ -15,7 +15,7 @@
 </script>
 
 <div class="root">
-  <Heading medium black>Start With a Template</Heading>
+  <Heading lh medium black>Start With a Template</Heading>
   {#await templatesPromise}
     <div class="spinner-container">
       <Spinner size="30" />
@@ -24,11 +24,12 @@
     <div class="templates">
       {#each templates as template}
         <div class="templates-card">
-          <Heading black medium>{template.name}</Heading>
+          <Heading black small>{template.name}</Heading>
+          <Spacer small />
           <Body medium grey>{template.category}</Body>
-          <Body small black>{template.description}</Body>
+          <Body lh small black>{template.description}</Body>
           <div>
-            <img src={template.image} width="300" />
+            <img src={template.image} width="100%" />
           </div>
           <div class="card-footer">
             <Button secondary on:click={() => onSelect(template)}>
@@ -67,9 +68,5 @@
     font-weight: 600;
     color: var(--ink);
     text-transform: capitalize;
-  }
-
-  .root {
-    margin: 20px 80px;
   }
 </style>
