@@ -1,6 +1,13 @@
 <script>
   import groupBy from "lodash/fp/groupBy"
-  import { Button, TextArea, Label, Body } from "@budibase/bbui"
+  import {
+    Button,
+    TextArea,
+    Label,
+    Body,
+    Heading,
+    Spacer,
+  } from "@budibase/bbui"
   import { createEventDispatcher } from "svelte"
 
   const dispatch = createEventDispatcher()
@@ -26,9 +33,10 @@
 
 <div class="container" data-cy="binding-dropdown-modal">
   <div class="list">
-    <Label size="l" color="dark">Objects</Label>
+    <Heading extraSmall>Objects</Heading>
+    <Spacer medium />
     {#if context}
-      <Label size="s" color="dark">Table</Label>
+      <Heading extraSmall>Tables</Heading>
       <ul>
         {#each context as { readableBinding }}
           <li on:click={() => addToText(readableBinding)}>{readableBinding}</li>
@@ -36,7 +44,7 @@
       </ul>
     {/if}
     {#if instance}
-      <Label size="s" color="dark">Components</Label>
+      <Heading extraSmall>Components</Heading>
       <ul>
         {#each instance as { readableBinding }}
           <li on:click={() => addToText(readableBinding)}>{readableBinding}</li>
@@ -45,15 +53,21 @@
     {/if}
   </div>
   <div class="text">
-    <Label size="l" color="dark">Data binding</Label>
-    <Body size="s" color="dark">
+    <Heading extraSmall>Data binding</Heading>
+    <Spacer small />
+    <Body extraSmall lh>
       Binding connects one piece of data to another and makes it dynamic. Click
       the objects on the left, to add them to the textbox.
     </Body>
-    <TextArea bind:value placeholder="" />
+    <Spacer large />
+    <TextArea
+      thin
+      bind:value
+      placeholder="Add text, or lick the objects on the left to add them to the
+      textbox." />
     <div class="controls">
-      <a href="#">
-        <Body size="s" color="light">Learn more about binding</Body>
+      <a href="https://docs.budibase.com/design/binding">
+        <Body small grey>Learn more about binding</Body>
       </a>
       <Button on:click={cancel} secondary>Cancel</Button>
       <Button on:click={close} primary>Done</Button>
@@ -73,17 +87,19 @@
   .controls {
     margin-top: var(--spacing-m);
     display: grid;
-    align-items: center;
+    align-items: baseline;
     grid-gap: var(--spacing-l);
     grid-template-columns: 1fr auto auto;
   }
   .list {
     width: 150px;
     border-right: 1.5px solid var(--grey-4);
+    padding: var(--spacing-xl);
   }
   .text {
     width: 600px;
-    display: grid;
+    padding: var(--spacing-xl);
+    font-family: var(--font-sans);
   }
   .text :global(p) {
     margin: 0;
@@ -99,15 +115,16 @@
     display: flex;
     font-family: var(--font-sans);
     font-size: var(--font-size-xs);
-    color: var(--ink);
-    padding: var(--spacing-s) var(--spacing-m);
+    color: var(--grey-7);
+    padding: var(--spacing-s) 0;
     margin: auto 0px;
     align-items: center;
     cursor: pointer;
   }
 
   li:hover {
-    background-color: var(--grey-2);
+    color: var(--ink);
+    font-weight: 500;
   }
 
   li:active {
