@@ -1,6 +1,12 @@
 <script>
   import { store } from "builderStore"
-  import { TextButton, Button, Heading, DropdownMenu } from "@budibase/bbui"
+  import {
+    TextButton,
+    Button,
+    Body,
+    DropdownMenu,
+    ModalContent,
+  } from "@budibase/bbui"
   import { AddIcon, ArrowDownIcon } from "components/common/Icons/"
   import { EVENT_TYPE_MEMBER_NAME } from "../../common/eventHandlers"
   import actionTypes from "./actions"
@@ -21,12 +27,6 @@
     selectedAction &&
     actionTypes.find(t => t.name === selectedAction[EVENT_TYPE_MEMBER_NAME])
       .component
-
-  const closeModal = () => {
-    dispatch("close")
-    draftEventHandler = { parameters: [] }
-    actions = []
-  }
 
   const updateEventHandler = (updatedHandler, index) => {
     actions[index] = updatedHandler
@@ -54,20 +54,17 @@
 
   const saveEventData = () => {
     dispatch("change", actions)
-    closeModal()
   }
 </script>
 
-<div class="root">
-
-  <div class="header">
-    <Heading small dark>Actions</Heading>
+<ModalContent title="Actions" confirmText="Save" onConfirm={saveEventData}>
+  <div slot="header">
     <div bind:this={addActionButton}>
       <TextButton text small blue on:click={addActionDropdown.show}>
-        Add Action
         <div style="height: 20px; width: 20px;">
           <AddIcon />
         </div>
+        Add Action
       </TextButton>
     </div>
     <DropdownMenu
@@ -89,11 +86,7 @@
       {#each actions as action, index}
         <div class="action-container">
           <div class="action-header" on:click={selectAction(action)}>
-            <p
-              class="bb-body bb-body--small bb-body--color-dark"
-              style="margin: var(--spacing-s) 0;">
-              {index + 1}. {action[EVENT_TYPE_MEMBER_NAME]}
-            </p>
+            <Body small lh>{index + 1}. {action[EVENT_TYPE_MEMBER_NAME]}</Body>
             <div class="row-expander" class:rotate={action !== selectedAction}>
               <ArrowDownIcon />
             </div>
@@ -115,30 +108,12 @@
     {/if}
   </div>
 
-  <div class="footer">
+  <div slot="footer">
     <a href="https://docs.budibase.com">Learn more about Actions</a>
-    <Button secondary on:click={closeModal}>Cancel</Button>
-    <Button primary on:click={saveEventData}>Save</Button>
   </div>
-</div>
+</ModalContent>
 
 <style>
-  .root {
-    max-height: 50vh;
-    width: 700px;
-    display: flex;
-    flex-direction: column;
-  }
-
-  .header {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-between;
-    padding: var(--spacing-xl);
-    padding-bottom: 0;
-  }
-
   .action-header {
     display: flex;
     flex-direction: row;
@@ -166,8 +141,7 @@
 
   .actions-container {
     flex: 1;
-    min-height: 0px;
-    padding-bottom: var(--spacing-s);
+    min-height: 0;
     padding-top: 0;
     border: var(--border-light);
     border-width: 0 0 1px 0;
@@ -177,10 +151,6 @@
   .action-container {
     border: var(--border-light);
     border-width: 1px 0 0 0;
-    padding-left: var(--spacing-xl);
-    padding-right: var(--spacing-xl);
-    padding-top: 0;
-    padding-bottom: 0;
   }
 
   .selected-action-container {
@@ -195,22 +165,13 @@
     flex-direction: row;
   }
 
-  .footer {
-    display: flex;
-    flex-direction: row;
-    gap: var(--spacing-s);
-    padding: var(--spacing-xl);
-    padding-top: var(--spacing-m);
-  }
-
-  .footer > a {
+  a {
     flex: 1;
     color: var(--grey-5);
     font-size: var(--font-size-s);
     text-decoration: none;
   }
-
-  .footer > a:hover {
+  a:hover {
     color: var(--blue);
   }
 
