@@ -1,25 +1,17 @@
 <script>
   import { Button, Modal } from "@budibase/bbui"
   import EventEditorModal from "./EventEditorModal.svelte"
-  import { createEventDispatcher, onMount } from "svelte"
+  import { createEventDispatcher } from "svelte"
   const dispatch = createEventDispatcher()
 
   export let value
   export let name
 
-  let eventsModal
+  let modal
 </script>
 
-<Button secondary small on:click={eventsModal.show}>Define Actions</Button>
+<Button secondary small on:click={modal.show}>Define Actions</Button>
 
-<Modal bind:this={eventsModal} maxWidth="100vw" hideCloseButton padding="0">
-  <EventEditorModal
-    event={value}
-    eventType={name}
-    on:change
-    on:close={eventsModal.hide} />
+<Modal bind:this={modal} width="600px">
+  <EventEditorModal event={value} eventType={name} on:change />
 </Modal>
-
-<style>
-
-</style>
