@@ -4,23 +4,21 @@
   import PageLayout from "./PageLayout.svelte"
   import PagesList from "./PagesList.svelte"
   import { store } from "builderStore"
-  import NewScreen from "./NewScreen.svelte"
-  import CurrentItemPreview from "./CurrentItemPreview.svelte"
+  import NewScreenModal from "./NewScreenModal.svelte"
+  import CurrentItemPreview from "./AppPreview/CurrentItemPreview.svelte"
   import SettingsView from "./SettingsView.svelte"
   import ComponentsPaneSwitcher from "./ComponentsPaneSwitcher.svelte"
   import ConfirmDialog from "components/common/ConfirmDialog.svelte"
   import { last } from "lodash/fp"
   import { AddIcon } from "components/common/Icons"
+  import { Modal } from "@budibase/bbui"
 
   let newScreenPicker
   let confirmDeleteDialog
   let componentToDelete = ""
-
-  const newScreen = () => {
-    newScreenPicker.show()
-  }
-
   let settingsView
+  let modal
+
   const settings = () => {
     settingsView.show()
   }
@@ -29,30 +27,24 @@
 </script>
 
 <div class="root">
-
   <div class="ui-nav">
-
     <div class="pages-list-container">
       <div class="nav-header">
         <span class="navigator-title">Navigator</span>
-
         <span class="components-nav-page">Pages</span>
       </div>
-
       <div class="nav-items-container">
         <PagesList />
       </div>
     </div>
-
     <PageLayout layout={$store.pages[$store.currentPageName]} />
-
     <div class="components-list-container">
       <div class="nav-group-header">
         <span class="components-nav-header" style="margin-top: 0;">
           Screens
         </span>
         <div>
-          <button on:click={newScreen}>
+          <button on:click={modal.show}>
             <AddIcon />
           </button>
         </div>
@@ -61,22 +53,21 @@
         <ComponentsHierarchy screens={$store.screens} />
       </div>
     </div>
-
   </div>
-
   <div class="preview-pane">
     <CurrentItemPreview />
   </div>
-
   {#if $store.currentFrontEndType === 'screen' || $store.currentFrontEndType === 'page'}
     <div class="components-pane">
       <ComponentsPaneSwitcher />
     </div>
   {/if}
-
 </div>
 
-<NewScreen bind:this={newScreenPicker} />
+<Modal bind:this={modal}>
+  <NewScreenModal />
+</Modal>
+
 <SettingsView bind:this={settingsView} />
 
 <style>
