@@ -1,12 +1,22 @@
 export const FIELDS = {
   STRING: {
-    name: "Plain Text",
+    name: "Text",
     icon: "ri-text",
     type: "string",
     constraints: {
       type: "string",
       length: {},
       presence: false,
+    },
+  },
+  OPTIONS: {
+    name: "Options",
+    icon: "ri-list-check-2",
+    type: "options",
+    constraints: {
+      type: "string",
+      presence: { allowEmpty: true },
+      inclusion: [],
     },
   },
   NUMBER: {
@@ -28,15 +38,6 @@ export const FIELDS = {
       presence: false,
     },
   },
-  // OPTIONS: {
-  //   name: "Options",
-  //   icon: "ri-list-check-2",
-  //   type: "options",
-  //   constraints: {
-  //     type: "string",
-  //     presence: { allowEmpty: true },
-  //   },
-  // },
   DATETIME: {
     name: "Date/Time",
     icon: "ri-calendar-event-fill",
@@ -60,15 +61,15 @@ export const FIELDS = {
       presence: false,
     },
   },
-  // LINKED_FIELDS: {
-  //   name: "Linked Fields",
-  //   icon: "ri-link",
-  //   type: "link",
-  //   modelId: null,
-  //   constraints: {
-  //     type: "array",
-  //   },
-  // },
+  LINK: {
+    name: "Relationship",
+    icon: "ri-link",
+    type: "link",
+    constraints: {
+      type: "array",
+      presence: { allowEmpty: true },
+    },
+  },
 }
 
 export const FILE_TYPES = {
