@@ -1,12 +1,6 @@
 <script>
-  import { getContext } from "svelte"
-  import { Button } from "@budibase/bbui"
-  import ModelDataTable from "components/database/DataTable"
+  import ModelDataTable from "components/backend/DataTable/ModelDataTable.svelte"
   import { backendUiStore } from "builderStore"
-  import ActionButton from "components/common/ActionButton.svelte"
-  import * as api from "components/database/DataTable/api"
-
-  const { open, close } = getContext("simple-modal")
 
   $: selectedModel = $backendUiStore.selectedModel
 </script>
@@ -14,12 +8,12 @@
 {#if $backendUiStore.selectedDatabase._id && selectedModel.name}
   <ModelDataTable />
 {:else}
-  <i style="color: var(--grey-4)">create your first table to start building</i>
+  <i>Create your first table to start building</i>
 {/if}
 
 <style>
   i {
-    font-size: 20px;
-    margin-right: 10px;
+    font-size: var(--font-size-xl);
+    color: var(--grey-4);
   }
 </style>
