@@ -21,19 +21,12 @@
 </script>
 
 <div class="root">
-  <div class="node-view">
-    <slot />
-  </div>
+  <slot />
 </div>
 
 <style>
   .root {
     height: 100%;
     position: relative;
-  }
-
-  .node-view {
-    overflow-y: auto;
-    flex: 1 1 auto;
   }
 </style>
