@@ -1,12 +1,6 @@
 <script>
-  import { getContext } from "svelte"
-  import { Button } from "@budibase/bbui"
-  import ViewDataTable from "components/database/DataTable/ViewDataTable"
+  import ViewDataTable from "components/backend/DataTable/ViewDataTable"
   import { backendUiStore } from "builderStore"
-  import ActionButton from "components/common/ActionButton.svelte"
-  import * as api from "components/database/DataTable/api"
-
-  const { open, close } = getContext("simple-modal")
 
   $: selectedView = $backendUiStore.selectedView
 </script>
@@ -14,13 +8,12 @@
 {#if $backendUiStore.selectedDatabase._id && selectedView}
   <ViewDataTable view={selectedView} />
 {:else}
-  <i>create your first table to start building</i>
+  <i>Create your first table to start building</i>
 {/if}
 
 <style>
   i {
-    font-size: 20px;
-    margin-right: 10px;
+    font-size: var(--font-size-xl);
     color: var(--grey-4);
   }
 </style>
