@@ -1,17 +1,31 @@
 <script>
-  import { getContext } from "svelte"
   import { store } from "builderStore"
   import api from "builderStore/api"
   import AppList from "components/start/AppList.svelte"
   import { onMount } from "svelte"
   import ActionButton from "components/common/ActionButton.svelte"
+  import { get } from "builderStore/api"
   import Spinner from "components/common/Spinner.svelte"
   import CreateAppModal from "components/start/CreateAppModal.svelte"
+  import { Button, Heading, Modal } from "@budibase/bbui"
   import TemplateList from "components/start/TemplateList.svelte"
-  import { Button } from "@budibase/bbui"
   import analytics from "analytics"
 
+  let promise = getApps()
   let hasKey
+  let template
+  let modal
+
+  async function getApps() {
+    const res = await get("/api/applications")
+    const json = await res.json()
+
+    if (res.ok) {
+      return json
+    } else {
+      throw new Error(json)
+    }
+  }
 
   async function fetchKeys() {
     const response = await api.get(`/api/keys/`)
@@ -27,62 +41,51 @@
     }
 
     if (!keys.budibase) {
-      showCreateAppModal()
+      modal.show()
     }
   }
 
-  // Handle create app modal
-  const { open } = getContext("simple-modal")
-
-  const showCreateAppModal = template => {
-    open(
-      CreateAppModal,
-      {
-        hasKey,
-        template,
-      },
-      {
-        closeButton: false,
-        closeOnEsc: false,
-        closeOnOuterClick: false,
-        styleContent: { padding: 0 },
-        closeOnOuterClick: true,
-      }
-    )
+  function selectTemplate(newTemplate) {
+    template = newTemplate
+    modal.show()
   }
 
   checkIfKeysAndApps()
 </script>
 
-<div class="header">
-  <div class="welcome">Welcome to the Budibase Beta</div>
-  <Button primary purple on:click={() => showCreateAppModal()}>
-    Create New Web App
-  </Button>
-</div>
-
-<div class="banner">
-  <img src="/_builder/assets/orange-landscape.png" alt="rocket" />
-  <div class="banner-content">
-    Every accomplishment starts with a decision to try.
+<div class="container">
+  <div class="header">
+    <Heading medium black>Welcome to the Budibase Beta</Heading>
+    <Button primary purple on:click={modal.show}>Create New Web App</Button>
   </div>
+
+  <div class="banner">
+    <img src="/_builder/assets/orange-landscape.png" alt="rocket" />
+    <div class="banner-content">
+      Every accomplishment starts with a decision to try.
+    </div>
+  </div>
+
+  <TemplateList onSelect={selectTemplate} />
+
+  <AppList />
 </div>
 
-<TemplateList onSelect={showCreateAppModal} />
-<AppList />
+<Modal bind:this={modal} padding={false} width="600px">
+  <CreateAppModal {hasKey} {template} />
+</Modal>
 
 <style>
+  .container {
+    display: grid;
+    gap: var(--spacing-xl);
+    margin: 40px 80px;
+  }
+
   .header {
     display: flex;
     justify-content: space-between;
     align-items: center;
-    margin: 40px 80px 0px 80px;
-  }
-
-  .welcome {
-    font-size: var(--font-size-3xl);
-    color: var(--ink);
-    font-weight: 600;
   }
 
   .banner {
@@ -92,7 +95,6 @@
     position: relative;
     text-align: center;
     color: white;
-    margin: 20px 80px 40px 80px;
     border-radius: 16px;
   }
 
