@@ -1,4 +1,5 @@
 const CouchDB = require("../../db")
+const linkRecords = require("../../db/linkedRecords")
 const csvParser = require("../../utilities/csvParser")
 const {
   getRecordParams,
@@ -23,7 +24,8 @@ exports.find = async function(ctx) {
 }
 
 exports.save = async function(ctx) {
-  const db = new CouchDB(ctx.user.instanceId)
+  const instanceId = ctx.user.instanceId
+  const db = new CouchDB(instanceId)
   const { dataImport, ...rest } = ctx.request.body
   const modelToSave = {
     type: "model",
@@ -31,15 +33,31 @@ exports.save = async function(ctx) {
     views: {},
     ...rest,
   }
+  // get the model in its previous state for differencing
+  let oldModel
+  let oldModelId = ctx.request.body._id
+  if (oldModelId) {
+    // if it errors then the oldModelId is invalid - can't diff it
+    try {
+      oldModel = await db.get(oldModelId)
+    } catch (err) {
+      oldModel = null
+    }
+  }
 
   // rename record fields when table column is renamed
   const { _rename } = modelToSave
-  if (_rename) {
+  if (_rename && modelToSave.schema[_rename.updated].type === "link") {
+    throw "Cannot rename a linked field."
+  } else if (_rename && modelToSave.primaryDisplay === _rename.old) {
+    throw "Cannot rename the primary display field."
+  } else if (_rename) {
     const records = await db.allDocs(
       getRecordParams(modelToSave._id, null, {
         include_docs: true,
       })
     )
+
     const docs = records.rows.map(({ doc }) => {
       doc[_rename.updated] = doc[_rename.old]
       delete doc[_rename.old]
@@ -62,23 +80,18 @@ exports.save = async function(ctx) {
   const result = await db.post(modelToSave)
   modelToSave._rev = result.rev
 
-  const { schema } = ctx.request.body
-  for (let key of Object.keys(schema)) {
-    // model has a linked record
-    if (schema[key].type === "link") {
-      // create the link field in the other model
-      const linkedModel = await db.get(schema[key].modelId)
-      linkedModel.schema[modelToSave.name] = {
-        name: modelToSave.name,
-        type: "link",
-        modelId: modelToSave._id,
-        constraints: {
-          type: "array",
-        },
-      }
-      await db.put(linkedModel)
-    }
-  }
+  // update linked records
+  await linkRecords.updateLinks({
+    instanceId,
+    eventType: oldModel
+      ? linkRecords.EventType.MODEL_UPDATED
+      : linkRecords.EventType.MODEL_SAVE,
+    model: modelToSave,
+    oldModel: oldModel,
+  })
+
+  ctx.eventEmitter &&
+    ctx.eventEmitter.emitModel(`model:save`, instanceId, modelToSave)
 
   if (dataImport && dataImport.path) {
     // Populate the table with records imported from CSV in a bulk update
@@ -98,7 +111,8 @@ exports.save = async function(ctx) {
 }
 
 exports.destroy = async function(ctx) {
-  const db = new CouchDB(ctx.user.instanceId)
+  const instanceId = ctx.user.instanceId
+  const db = new CouchDB(instanceId)
 
   const modelToDelete = await db.get(ctx.params.modelId)
 
@@ -114,16 +128,15 @@ exports.destroy = async function(ctx) {
     records.rows.map(record => ({ _id: record.id, _deleted: true }))
   )
 
-  // Delete linked record fields in dependent models
-  for (let key of Object.keys(modelToDelete.schema)) {
-    const { type, modelId } = modelToDelete.schema[key]
-    if (type === "link") {
-      const linkedModel = await db.get(modelId)
-      delete linkedModel.schema[modelToDelete.name]
-      await db.put(linkedModel)
-    }
-  }
+  // update linked records
+  await linkRecords.updateLinks({
+    instanceId,
+    eventType: linkRecords.EventType.MODEL_DELETE,
+    model: modelToDelete,
+  })
 
+  ctx.eventEmitter &&
+    ctx.eventEmitter.emitModel(`model:delete`, instanceId, modelToDelete)
   ctx.status = 200
   ctx.message = `Model ${ctx.params.modelId} deleted.`
 }
